+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 ### Changed
 
 - Network details panel now opens as a new tab in the network tool window.
+- Move icons to left side of network list window.
 
 ### Added